+++ arch/x86/lib/csum-wrappers_64.c
@@ -5,8 +5,7 @@
  * Wrappers of assembly checksum functions for x86-64.
  */
 #include <asm/checksum.h>
-#include <linux/export.h>
-#include <linux/uaccess.h>
+#include <linux/module.h>
 #include <asm/smap.h>
 
 /**
@@ -136,7 +135,7 @@ EXPORT_SYMBOL(csum_partial_copy_nocheck);
 
 __sum16 csum_ipv6_magic(const struct in6_addr *saddr,
 			const struct in6_addr *daddr,
-			__u32 len, __u8 proto, __wsum sum)
+			__u32 len, unsigned short proto, __wsum sum)
 {
 	__u64 rest, sum64;
